requirements.txt
torch
torchvision
torchaudio
sentence_transformers
numpy
faiss-cpu
networkx
pandas
scipy
tqdm
python-louvain
spacy
streamlit